requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==23.2.0
calplot==0.1.7.5
chardet==4.0.0
contourpy==1.2.0
cycler==0.12.1
DataProperty==1.0.1
discord.py==2.3.2
exceptiongroup==1.2.0
fonttools==4.47.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
install==1.3.5
kiwisolver==1.4.5
matplotlib==3.8.2
mbstrdecoder==1.1.3
multidict==6.0.4
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pathvalidate==3.2.0
pillow==10.2.0
pluggy==1.3.0
pyparsing==3.1.1
pytablewriter==1.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-md-report==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
table2ascii==1.1.3
tabledata==1.3.3
tcolorpy==0.1.4
tomli==2.0.1
typepy==1.3.2
typing_extensions==4.9.0
tzdata==2023.4
wcwidth==0.2.13
yarl==1.9.4
